requirements.txt
psycopg2
pandas
requests
google-cloud-bigquery
google-auth
python-dotenv
apscheduler
db-dtypes
pytz